Fixed categories time

+++ shooting-ipsc/src/main/java/tech/shooting/ipsc/bean/CategoriesBean.java
@@ -21,15 +21,15 @@ public class CategoriesBean {
     @JsonProperty
     @ApiModelProperty(value = "Excellent execute time for this category", required = true)
     @Positive(message = ValidationConstants.TIME_MESSAGE)
-    private Long excellentTime;
+    private Float excellentTime;
     
     @JsonProperty
     @ApiModelProperty(value = "Good execute time for this category", required = true)
     @Positive(message = ValidationConstants.TIME_MESSAGE)
-    private Long goodTime;
+    private Float goodTime;
     
     @JsonProperty("salTime")
     @ApiModelProperty(value = "Standard execute time for this category", required = true)
     @Positive(message = ValidationConstants.TIME_MESSAGE)
-    private Long salTime;
+    private Float salTime;
 }
